requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
requests>=2.32.0,<2.32.3
pydantic>=2.10.0,<2.10.4